requirements.txt
pipecat-ai[daily,elevenlabs,openai,silero,google]
fastapi
uvicorn
python-dotenv
twilio
python-multipart
aiohttp
